+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Toolium Changelog
 v3.0.2
 ------
 
-*Release date: In development*
+*Release date: 2023-06-08*
 
 - `context.storage` must be initialized before dynamic environment steps in `before_feature` method
 - Mark scenario as failed when a dynamic environment step fails in `before_scenario`